refactor(rts-core): delete CodebaseAnalyzer + sibling types (B3)
Architecture-strategist's "kill the bigger thing" decision from the
plan's Decisions Applied (S2). After PR-A stripped FileCache +
SemanticGraphQuery from `analyzer.rs` and B2 migrated the daemon's
sole consumer to `parse_content`, `CodebaseAnalyzer` was unamortized
scaffolding around a parse-and-extract call the daemon already made
per-invocation.

Deleted:
- `crates/rts-core/src/analyzer.rs` (1607 LoC)
- `pub use analyzer::{AnalysisConfig, AnalysisDepth, AnalysisResult,
  CodebaseAnalyzer, FileInfo, Symbol}` from lib.rs

Added:
- `crates/rts-core/src/symbol.rs` (33 LoC) — the `Symbol` payload's
  new home, next to its only producer (`extraction.rs`). Re-exported
  from lib.rs as before; daemon's `use rust_tree_sitter::Symbol`
  works unchanged.
- 15 extraction tests preserved in `extraction::tests` — rewritten
  to use `parse_content` (the public API) instead of
  `CodebaseAnalyzer::analyze_directory`. Coverage of go/js/jsdoc/rust-
  const+static/ruby/java/swift/csharp/php (interface/trait/namespaced)
  doc-comment extraction + Rust trait/type/union/macro extraction is
  fully preserved against the public surface.
- `parse_content_extracts_multiple_rust_kinds` replaces the B1-era
  equivalence test (which referenced `CodebaseAnalyzer`).

Grep `CodebaseAnalyzer` across the workspace post-commit: zero hits
in `*.rs` source files (changelog/plan/doc mentions retained as
historical record).

`cargo test -p rust_tree_sitter --lib`: 201/201 pass.

B4 (rewriting / deleting the 4 integration tests that imported the
deleted types) follows in subsequent commits.
